🐛 fix(VSecM): fix failing tests
